requirements.txt
playwright>=1.40.0
google-genai
Pillow
Flask
python-dotenv
